+++ ui/src/pages/SignInPage/SignInPage.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 
-import Layout from '../components/Layout';
+import Layout from '../../components/Layout';
 import { useNavigate } from 'react-router';
 
 interface SignInPageProps {
